Fixed issue where captured-via-app photo file was being deleted, in edge cases where it shouldn't be

+++ iNaturalist/src/main/java/org/inaturalist/android/ObservationEditor.java
@@ -3258,9 +3258,11 @@ public class ObservationEditor extends Fragment {
             mCameraPhotos.add(copyPath);
         }
 
-        // Delete original photo (before resize)
-        File f = new File(path);
-        f.delete();
+        if ((copyPath != null) && (!copyPath.equals(path))) {
+            // Delete original photo (before resize)
+            File f = new File(path);
+            f.delete();
+        }
 
         if (createdUri == null) {
             getActivity().runOnUiThread(new Runnable() {
